prov/net: Convert ssize_t return type to int
The net provider was initially using ssize_t return types because
the ofi_bsock_* functions were returning ssize_t values as well.

Since the bsock interface now exclusively uses int for the return type,
we can convert all the ssize_t return types from the provider to int.

+++ prov/net/src/xnet.h
@@ -157,7 +157,7 @@ struct xnet_active_rx {
 	size_t			hdr_done;
 	size_t			data_left;
 	struct xnet_xfer_entry	*entry;
-	ssize_t			(*handler)(struct xnet_ep *ep);
+	int			(*handler)(struct xnet_ep *ep);
 	void			*claim_ctx;
 };
 
@@ -323,7 +323,7 @@ int xnet_init_progress(struct xnet_progress *progress, struct fi_info *info);
 void xnet_close_progress(struct xnet_progress *progress);
 int xnet_start_progress(struct xnet_progress *progress);
 void xnet_stop_progress(struct xnet_progress *progress);
-ssize_t xnet_start_recv(struct xnet_ep *ep, struct xnet_xfer_entry *rx_entry);
+int xnet_start_recv(struct xnet_ep *ep, struct xnet_xfer_entry *rx_entry);
 
 void xnet_progress(struct xnet_progress *progress, bool clear_signal);
 void xnet_run_progress(struct xnet_progress *progress, bool clear_signal);
